:recycle: refactor(tofhir-server-common): Move ICORSHandler of tofhir-log-server and tofhir-server.
