+++ tofhir-server/src/main/scala/io/tofhir/server/endpoint/JobEndpoint.scala
@@ -12,7 +12,7 @@ import io.tofhir.server.model.{ExecuteJobTask, RowSelectionOrder, TestResourceCr
 import io.tofhir.server.service.{ExecutionService, JobService}
 import io.tofhir.engine.Execution.actorSystem.dispatcher
 import io.tofhir.engine.util.FhirMappingJobFormatter.formats
-import io.tofhir.server.interceptor.ICORSHandler
+import io.tofhir.server.common.interceptor.ICORSHandler
 import io.tofhir.server.service.job.IJobRepository
 import io.tofhir.server.service.mapping.IMappingRepository
 import io.tofhir.server.service.schema.ISchemaRepository
