+++ tofhir-server/src/main/scala/io/tofhir/server/endpoint/ToFhirServerEndpoint.scala
@@ -7,7 +7,8 @@ import io.tofhir.engine.config.ToFhirEngineConfig
 import io.tofhir.server.config.LogServiceConfig
 import io.tofhir.server.common.config.WebServerConfig
 import io.tofhir.server.fhir.FhirDefinitionsConfig
-import io.tofhir.server.interceptor.{ICORSHandler, IErrorHandler}
+import io.tofhir.server.interceptor.IErrorHandler
+import io.tofhir.server.common.interceptor.ICORSHandler
 import io.tofhir.server.model.ToFhirRestCall
 import io.tofhir.server.service.job.JobFolderRepository
 import io.tofhir.server.service.mapping.ProjectMappingFolderRepository
