+++ tofhir-server/src/main/scala/io/tofhir/server/util/LogServiceClient.scala
@@ -4,7 +4,7 @@ import akka.http.scaladsl.Http
 import akka.http.scaladsl.model.{HttpMethods, HttpRequest}
 import io.tofhir.engine.Execution.actorSystem
 import io.tofhir.engine.Execution.actorSystem.dispatcher
-import io.tofhir.server.interceptor.ICORSHandler
+import io.tofhir.server.common.interceptor.ICORSHandler
 import io.tofhir.server.model.Json4sSupport.formats
 import org.json4s.JValue
 import org.json4s.JsonAST.JObject
